pyproject.toml
[tool.poetry]
name = "redpillinvest"
version = "0.1.0"
description = ""
authors = ["Luca Knaack <l.knaack93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
django-filter = "^24.3"
celery = "^5.4.0"
redis = "^5.0.8"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
